# api/openapi/aor.v1.yaml
openapi: 3.2.0
info:
  title: Agent Organization Runtime API
  version: 1.0.0
  description: Deterministic, tenant-scoped control API for SPEC 2.0.0.
servers:
  - url: https://aor.example.invalid
security:
  - oidc: [aor]
  - mutualTLS: []
...
paths:
  /v1/model-providers:
    get:
      tags: [Settings]
      operationId: listModelProviders
      responses:
        '200':
          description: Configured provider metadata without endpoints or credential references
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ModelProviderPage'}
        default: {$ref: '#/components/responses/Problem'}
  /v1/settings/model-providers:
    get:
      tags: [Settings]
      operationId: getModelProviderSettings
      responses:
        '200':
          description: Provider endpoints and write-only credential status
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ModelProviderSettingsPage'}
        default: {$ref: '#/components/responses/Problem'}
  /v1/settings/model-providers/{providerId}:
    parameters:
      - $ref: '#/components/parameters/ProviderId'
    put:
      tags: [Settings]
      operationId: putModelProviderSettings
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ModelProviderSettingsInput'}
      responses:
        '200':
          description: Saved provider settings without the API key
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ModelProviderSettings'}
        default: {$ref: '#/components/responses/Problem'}
  /v1/settings/model-providers/{providerId}:test:
    parameters:
      - $ref: '#/components/parameters/ProviderId'
    post:
      tags: [Settings]
      operationId: testModelProvider
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ModelProviderTestInput'}
      responses:
        '200':
          description: Sanitized connectivity result
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ModelProviderTestResult'}
        default: {$ref: '#/components/responses/Problem'}
...
  /v1/settings/model-sampling:
    get:
      tags: [Settings]
      operationId: getModelSamplingSettings
      responses:
        '200':
          description: Tenant-global generation parameters used for subsequent model calls
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ModelSamplingSettings'}
        default: {$ref: '#/components/responses/Problem'}
    put:
      tags: [Settings]
      operationId: putModelSamplingSettings
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ModelSamplingSettingsInput'}
      responses:
        '200':
          description: Saved tenant-global generation parameters
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ModelSamplingSettings'}
        default: {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.example.invalid/.well-known/openid-configuration
    mutualTLS:
      type: mutualTLS
...
  parameters:
...
    ProviderId: {name: providerId, in: path, required: true, schema: {enum: [openai, deepseek, claude, grok]}}
...
  schemas:
...
    ModelSamplingSettingsInput:
      type: object
      additionalProperties: false
      required: [temperature, topP, topK]
      properties:
        temperature: {type: number, minimum: 0, maximum: 2}
        topP: {type: number, minimum: 0, maximum: 1}
        topK: {type: integer, minimum: 0, maximum: 500}
    ModelSamplingSettings:
      type: object
      additionalProperties: false
      required: [temperature, topP, topK, version]
      properties:
        temperature: {type: number, minimum: 0, maximum: 2}
        topP: {type: number, minimum: 0, maximum: 1}
        topK: {type: integer, minimum: 0, maximum: 500}
        version: {type: integer, format: int64, minimum: 0}
    ModelProvider:
      type: object
      additionalProperties: false
      required: [id, provider, models, inputMicrosPerToken, outputMicrosPerToken, supportsStreaming, supportsToolCalls, supportsJsonSchema, supportsSeed, supportsPromptCaching, maxInputTokens, maxOutputTokens, allowedDataClassifications, dataResidency, retentionPolicy, modalities]
      properties:
        id: {type: string, minLength: 1, maxLength: 128}
        provider: {type: string, minLength: 1, maxLength: 128}
        models:
          type: array
          minItems: 1
          uniqueItems: true
          items: {type: string, minLength: 1, maxLength: 256}
        inputMicrosPerToken: {type: integer, format: int64, minimum: 0}
        outputMicrosPerToken: {type: integer, format: int64, minimum: 0}
        supportsStreaming: {type: boolean}
        supportsToolCalls: {type: boolean}
        supportsJsonSchema: {type: boolean}
        supportsSeed: {type: boolean}
        supportsPromptCaching: {type: boolean}
        maxInputTokens: {type: integer, minimum: 1}
        maxOutputTokens: {type: integer, minimum: 1}
        allowedDataClassifications:
          type: array
          minItems: 1
          uniqueItems: true
          items: {type: string, minLength: 1, maxLength: 64}
        dataResidency:
          type: array
          minItems: 1
          uniqueItems: true
          items: {type: string, minLength: 1, maxLength: 128}
        retentionPolicy: {type: string, minLength: 1, maxLength: 256}
        modalities:
          type: array
          minItems: 1
          uniqueItems: true
          items: {type: string, minLength: 1, maxLength: 128}
    ModelProviderPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/ModelProvider'}
    ModelProviderSettings:
      type: object
      additionalProperties: false
      required: [id, provider, displayName, baseUrl, protocol, protocols, models, apiKeyConfigured, enabled, version]
      properties:
        id: {enum: [openai, deepseek, claude, grok]}
        provider: {enum: [openai, deepseek, claude, grok]}
        displayName: {type: string, minLength: 1, maxLength: 64}
        baseUrl: {type: string, maxLength: 2048}
        protocol: {enum: [openai-compatible, anthropic-messages]}
        protocols:
          type: array
          minItems: 1
          uniqueItems: true
          items: {enum: [openai-compatible, anthropic-messages]}
        models:
          type: array
          minItems: 1
          uniqueItems: true
          items: {type: string, minLength: 1, maxLength: 256}
        apiKeyConfigured: {type: boolean}
        enabled: {type: boolean}
        version: {type: integer, format: int64, minimum: 0}
    ModelProviderSettingsPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          minItems: 4
          maxItems: 4
          items: {$ref: '#/components/schemas/ModelProviderSettings'}
    ModelProviderSettingsInput:
      type: object
      additionalProperties: false
      required: [baseUrl, protocol, enabled]
      properties:
        baseUrl: {type: string, maxLength: 2048}
        protocol: {enum: [openai-compatible, anthropic-messages]}
        apiKey: {type: string, minLength: 1, maxLength: 65536, writeOnly: true}
        enabled: {type: boolean}
    ModelProviderTestInput:
      type: object
      additionalProperties: false
      required: [baseUrl, protocol, model, reasoningEffort]
      properties:
        baseUrl: {type: string, minLength: 1, maxLength: 2048}
        protocol: {enum: [openai-compatible, anthropic-messages]}
        apiKey: {type: string, minLength: 1, maxLength: 65536, writeOnly: true}
        model: {type: string, minLength: 1, maxLength: 256}
        reasoningEffort: {enum: ['', none, low, medium, high, xhigh, max]}
    ModelProviderTestResult:
      type: object
      additionalProperties: false
      required: [ok, model]
      properties:
        ok: {type: boolean}
        model: {type: string, minLength: 1, maxLength: 256}
        latencyMs: {type: integer, format: int64, minimum: 0}
        detail: {type: string, maxLength: 256}
...
  responses:
...
    Problem:
      description: Redacted AOR Problem Details
      content:
        application/problem+json:
          schema: {$ref: '../json-schema/aor-error.v1.schema.json'}
...